feat: move bridge h3 regression tests to tests folder
